Add student update functionality and fixed StudentDto datatype bug

+++ student-Management/src/main/java/com/poorna/student/Management/service/StudentService.java
@@ -1,17 +1,17 @@
 package com.poorna.student.Management.service;
-
+import com.poorna.student.Management.Dto.StudentDto;
+import com.poorna.student.Management.Mapper.StudentMapper;
 import com.poorna.student.Management.model.Student;
 import com.poorna.student.Management.repository.StudentRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
 public class StudentService {
 
     private final StudentRepository studentRepository;
-
-
     public StudentService(StudentRepository studentRepository) {
         this.studentRepository = studentRepository;
     }
@@ -20,12 +20,37 @@ public class StudentService {
         return studentRepository.save(student);
     }
 
-    public Optional<Student> getStudentById(long id) {
-        return studentRepository.findById(id);
-    }
-    public String deleteStudentById(long id){
-         studentRepository.deleteById(id);
-        return "Deleted successfully";
+    public StudentDto getStudentById(long id) {
+        Student student = studentRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Student not found"));
+        return StudentMapper.toStudentDto(student);
     }
 
+    public String deleteStudentById(long id){
+         if(!studentRepository.existsById(id)){
+             throw new RuntimeException("Student not found");
+         }
+         studentRepository.deleteById(id);
+         return "Student ("+ id + ") deleted.";
+    }
+
+    public StudentDto UpdateStudent(Long id, Student updatedStudent){
+
+        Student existingStudent  = studentRepository.findById(id).
+                orElseThrow(() -> new RuntimeException("Student not found"));
+
+        existingStudent.setName(updatedStudent.getName());
+        existingStudent.setAge(updatedStudent.getAge());
+        existingStudent.setDepartment(updatedStudent.getDepartment());
+        existingStudent.setDateOfBirth(updatedStudent.getDateOfBirth());
+        existingStudent.setGender(updatedStudent.getGender());
+        existingStudent.setMobileNum(updatedStudent.getMobileNum());
+        existingStudent.setEmail(updatedStudent.getEmail());
+
+        return  StudentMapper.toStudentDto(studentRepository.save(existingStudent));
+    }
+
+    public List<Student> getAllStudents(){
+        return studentRepository.findAll();
+    }
 }
